tcp-reassembly.py: add utility to investigate TCP reassembly issues
While working on improving handling of reassembly in presence of
retransmissions, it would be very helpful to have a tool that tells the
correct interpretation. This tool does that. It can probably not
directly be implemented in Wireshark due to the additional memory
requirements.

Used to investigate bugs such as
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=13523
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=13061