sslkeylog.c: fix OpenSSL 1.1.1 support when library is not loaded
When libssl.so.1.1.1 was not yet loaded, it would assume that an older
library was already loaded and avoid the new API. That assumption is not
correct, it is also possible that no library was loaded at all as is the
case with Python. Test:

    ./sslkeylog.sh python -c \
        'import requests;print(requests.head("https://wireshark.org"))'

Before this fix it would output all zeroes as secret (a sign that
something is wrong).